+++ pay/docs.go
@@ -1,0 +1,4 @@
+/*
+Package pay contains the client-side functionality for easy use of APIs that are paywalled with ln-paywall or other compatible paywall implementations.
+*/
+package pay